removed prop drilling and used context
