+++ src/app/containers/MainContainer.jsx
@@ -1,30 +1,16 @@
-import React, { useReducer, useEffect } from 'react';
+import React, { useEffect } from 'react';
 import HeadContainer from './HeadContainer';
 import ActionContainer from './ActionContainer';
 import StateContainer from './StateContainer';
 import TravelContainer from './TravelContainer';
 import ButtonsContainer from './ButtonsContainer';
 
-import mainReducer from '../reducers/mainReducer';
 import { addNewSnapshots, initialConnect, setPort } from '../actions/actions';
-import StoreContext from '../store';
+import { useStoreContext } from '../store';
 
-const initialState = {
-  port: null,
-  sliderIndex: 0,
-  viewIndex: -1,
-  intervalId: null,
-  playing: false,
-  snapshots: [],
-  mode: {
-    locked: false,
-    paused: false,
-    persist: false,
-  },
-};
 
 function MainContainer() {
-  const [mainState, dispatch] = useReducer(mainReducer, initialState);
+  const [mainState, dispatch] = useStoreContext();
 
   useEffect(() => {
     if (mainState.port) return;
@@ -63,38 +49,20 @@ function MainContainer() {
     snapshots,
     sliderIndex,
     viewIndex,
-    playing,
-    mode,
   } = mainState;
 
   // if viewIndex is -1, then use the sliderIndex instead
   const snapshotView = (viewIndex === -1) ? snapshots[sliderIndex] : snapshots[viewIndex];
   return (
-    <StoreContext.Provider value={[mainState, dispatch]}>
-      <div className="main-container">
-        <HeadContainer />
-        <div className="body-container">
-          <ActionContainer
-            snapshots={snapshots}
-            sliderIndex={sliderIndex}
-            viewIndex={viewIndex}
-            dispatch={dispatch}
-          />
-          {(snapshots.length) ? <StateContainer snapshot={snapshotView} /> : null}
-          <TravelContainer
-            snapshotsLength={snapshots.length}
-            sliderIndex={sliderIndex}
-            playing={playing}
-            dispatch={dispatch}
-          />
-          <ButtonsContainer
-            mode={mode}
-            dispatch={dispatch}
-            snapshots={mainState.snapshots}
-          />
-        </div>
+    <div className="main-container">
+      <HeadContainer />
+      <div className="body-container">
+        <ActionContainer />
+        {(snapshots.length) ? <StateContainer snapshot={snapshotView} /> : null}
+        <TravelContainer snapshotsLength={snapshots.length} />
+        <ButtonsContainer />
       </div>
-    </StoreContext.Provider>
+    </div>
   );
 }
